requirements.txt
aiohttp
python-socketio
python-engineio==3.13.1
toml
pigpio